+++ Revo.Infrastructure/Security/Commands/CommandPermissionCache.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Revo.Core.Collections;
 using Revo.Core.Commands;
-using Revo.Core.Core;
 using Revo.Core.Lifecycle;
 using Revo.Core.Security;
 using Revo.Core.Types;
@@ -13,17 +11,17 @@ namespace Revo.Infrastructure.Security.Commands
     public class CommandPermissionCache : ICommandPermissionCache, IApplicationStartedListener
     {
         private readonly IPermissionTypeRegistry permissionTypeRegistry;
-        private readonly PermissionTypeIndexer permissionTypeIndexer;
-        private readonly ITypeExplorer typeExplorer;
+        private readonly IPermissionTypeIndexer permissionTypeIndexer;
+        private readonly ICommandTypeDiscovery commandTypeDiscovery;
         private Lazy<Dictionary<Type, CommandTypeInfo>> commandTypePermissions;
 
         public CommandPermissionCache(IPermissionTypeRegistry permissionTypeRegistry,
-            PermissionTypeIndexer permissionTypeIndexer,
-            ITypeExplorer typeExplorer)
+            IPermissionTypeIndexer permissionTypeIndexer,
+            ICommandTypeDiscovery commandTypeDiscovery)
         {
             this.permissionTypeRegistry = permissionTypeRegistry;
             this.permissionTypeIndexer = permissionTypeIndexer;
-            this.typeExplorer = typeExplorer;
+            this.commandTypeDiscovery = commandTypeDiscovery;
 
             Clear();
         }
@@ -35,7 +33,8 @@ namespace Revo.Infrastructure.Security.Commands
 
         public IReadOnlyCollection<Permission> GetCommandPermissions(ICommandBase command)
         {
-            if (commandTypePermissions.Value.TryGetValue(command.GetType(), out var info))
+            var commandType = GetCommandType(command);
+            if (commandTypePermissions.Value.TryGetValue(commandType, out var info))
             {
                 return info.Permissions;
             }
@@ -46,7 +45,8 @@ namespace Revo.Infrastructure.Security.Commands
 
         public bool IsAuthenticationRequired(ICommandBase command)
         {
-            if (commandTypePermissions.Value.TryGetValue(command.GetType(), out var info))
+            var commandType = GetCommandType(command);
+            if (commandTypePermissions.Value.TryGetValue(commandType, out var info))
             {
                 return info.IsAuthenticationRequired;
             }
@@ -62,10 +62,7 @@ namespace Revo.Infrastructure.Security.Commands
                 var dict = new Dictionary<Type, CommandTypeInfo>();
                 permissionTypeIndexer.EnsureIndexed();
 
-                foreach (var commandType in typeExplorer
-                    .GetAllTypes()
-                    .Where(x => x.IsClass && !x.IsAbstract && !x.IsGenericTypeDefinition)
-                    .Where(x => typeof(ICommandBase).IsAssignableFrom(x)))
+                foreach (var commandType in commandTypeDiscovery.DiscoverCommandTypes())
                 {
                     var permissions = GetCommandTypePermissions(commandType).ToArray();
                     bool isAuthenticated = IsAuthenticationRequired(commandType);
@@ -74,6 +71,17 @@ namespace Revo.Infrastructure.Security.Commands
 
                 return dict;
             });
+        }
+
+        private Type GetCommandType(ICommandBase command)
+        {
+            var commandType = command.GetType();
+            if (commandType.IsConstructedGenericType)
+            {
+                commandType = commandType.GetGenericTypeDefinition();
+            }
+
+            return commandType;
         }
 
         private IEnumerable<Permission> GetCommandTypePermissions(Type type)